requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-SSLify==0.1.5
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.7.0
ProxyTypes==0.9
WTForms==2.0.1
Werkzeug==0.9.6
argparse==1.2.1
blinker==1.3
docopt==0.4.0
flask-mongoengine==0.7.1
flufl.enum==4.0.1
gunicorn==19.1.1
httplib2==0.9
itsdangerous==0.24
mongoengine==0.8.7
py-bcrypt==0.4
pymongo==2.7.2
python-dateutil==2.4.0
python-loaders==0.2.3
pytz==2015.2
requests==2.6.0
six==1.9.0
wsgiref==0.1.2
